LAB-8 (Home [Unfinished])
It's 3 AM in the morning, my eyes are burning! I can't even anticipate a code when it actually works, I just have to move on to the next. Everytime I look at the ocean of files I need to write, I feel like I'm going nowhere with these. Why are we doing this in one day while this was supposed to be spread out throughout the entire semester?