LAB-6 (Sample)
MS Word is a document editor, not a code editor. It doesn't correct your mistakes. Please consider using a text editor for this purpose. And please for the love of God, compile the programs you're writing to check for syntax errors at least. I'm not even talking about Segfaults yet.